+++ src/DependencyInjection/Configuration.php
@@ -5,16 +5,8 @@ namespace Rokka\RokkaClientBundle\DependencyInjection;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
 
-/**
- * Class Configuration.
- */
 class Configuration implements ConfigurationInterface
 {
-    /**
-     * Generates the configuration tree builder.
-     *
-     * @return TreeBuilder The tree builder
-     */
     public function getConfigTreeBuilder()
     {
         $treeBuilder = new TreeBuilder();
@@ -27,10 +19,6 @@ class Configuration implements ConfigurationInterface
                     ->info('Api-Key value given on registration of the user.')
                     ->isRequired()
                 ->end()
-                ->scalarNode('api_secret')
-                    ->info('Api-secret value from user registration.')
-                    ->isRequired()
-                ->end()
                 ->scalarNode('organization')
                     ->info('Organization value like "my-org" for use in calls to the API.')
                     ->isRequired()
@@ -40,6 +28,14 @@ class Configuration implements ConfigurationInterface
                     ->defaultValue('https://api.rokka.io')
                 ->end()
             ->end();
+
+        $secretNode = $rootNode->children()->scalarNode('api_secret')
+            ->info('Deprecated. This is not used anymore.')
+            ->defaultNull();
+
+        if (method_exists($secretNode, 'setDeprecated')) {
+            $secretNode->setDeprecated();
+        }
 
         return $treeBuilder;
     }