+++ tests/Functional/Fixtures/AppKernel.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Rokka\RokkaClientBundle\Tests\Functional\Fixtures;
+
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\HttpKernel\Kernel;
 
@@ -15,7 +17,7 @@ class AppKernel extends Kernel
 
     public function registerContainerConfiguration(LoaderInterface $loader)
     {
-        $loader->load(__DIR__.'/config/config.yml');
+        $loader->load(__DIR__.'/app/config/config.yml');
     }
 
     public function getCacheDir()
@@ -28,3 +30,6 @@ class AppKernel extends Kernel
         return sys_get_temp_dir().'/rokka-client-bundle/logs';
     }
 }
+
+/* BC to run tests on Symfony < 3.3 */
+class_alias(AppKernel::class, \AppKernel::class, false);